+++ src/App.tsx
@@ -5,11 +5,11 @@ import Error404 from './components/Error404/Error404';
 import Tutorial from './Modules/Tutorial/Tutorial';
 import StartPage from './Modules/StartPage/StartPage';
 import GameBoard from './Modules/GameBoard/GameBoard';
+import OptionsPage from './Modules/OptionsPage/OptionsPage';
 import StartGameOption from './Modules/StartGameOption/StartGameOption';
 import NotifyComponent from './components/NotifyComponent/NotifyComponent';
 import { useAppSelector } from './app/hooks';
 import './App.scss';
-import OptionsPage from './Modules/OptionsPage/OptionsPage';
 
 function App() {
   const theme = useAppSelector((state) => state.theme);
