+++ src/Modules/OptionsPage/OptionsPage.scss
@@ -1,0 +1,3 @@
+.options-page{
+  
+}
