+++ src/Modules/OptionsPage/OptionsPage.tsx
@@ -1,8 +1,9 @@
 import './OptionsPage.scss';
+import BurgerMenu from '../../components/BurgerMenu/BurgerMenu';
 
 const OptionsPage = () => (
-    <div className="options-page">
-      <h1>heelo</h1>
+    <div className='options-page'>
+      <BurgerMenu/>
     </div>
 );
