+++ src/components/BurgerMenu/BurgerMenu.scss
@@ -1,0 +1,3 @@
+.burger-menu{
+  height: 440px;
+}
